test: wip and rename files
